Добавит директории в соответствии с критериями.
