+++ src/components/city-list/city.tsx
@@ -1,6 +1,6 @@
-import { CityName } from '../const';
-import { useAppDispatch } from '../hooks';
-import { offerData } from '../store/offer-data/offer-data';
+import { CityName } from '../../const';
+import { useAppDispatch } from '../../hooks';
+import { offerData } from '../../store/offer-data/offer-data';
 
 type CityProps = {
   cityName: CityName;
