+++ src/components/favorite-list/favorite-city.tsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
-import { CityName, Path } from '../const';
-import { useAppSelector } from '../hooks';
-import { getFavoriteOffers } from '../store/offer-data/offer-data-selectors';
+import { CityName, Path } from '../../const';
+import { useAppSelector } from '../../hooks';
+import { getFavoriteOffers } from '../../store/offer-data/offer-data-selectors';
 import FavoriteCard from './favorite-card';
 
 type FavoriteCityProps = {
