+++ src/components/no-favorite-offers/no-favorite-offers.tsx
