+++ src/components/offer-good-item/offer-good-item.tsx
