+++ src/components/offer-picture/offer-picture.tsx
