+++ src/pages/cities-list-screen/cities-list-screen.tsx
@@ -5,10 +5,10 @@ import { store } from '../../store';
 import { offerData } from '../../store/offer-data/offer-data';
 import { useEffect } from 'react';
 import { CityName } from '../../const';
-import CityList from '../../components/city-list';
-import Header from '../../components/header';
-import NoCityOffers from '../../components/no-city-offers';
-import CityOffersList from '../../components/city-offers-list';
+import CityList from '../../components/city-list/city-list';
+import Header from '../../components/header/header';
+import NoCityOffers from '../../components/no-city-offers/no-city-offers';
+import CityOffersList from '../../components/city-offers-list/city-offers-list';
 
 function CitiesListScreen (): JSX.Element {
   const city = useAppSelector(getCity);
