+++ src/store/index.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { createAPI } from '../services/api';
-import { rootReducer } from './root-reducer';
+import { rootReducer } from './root-reducer/root-reducer';
 
 const api = createAPI();
