+++ src/store/root-reducer/root-reducer.ts
@@ -1,7 +1,7 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import { NameSpace } from '../const';
-import { offerData } from './offer-data/offer-data';
-import { userProcess } from './user-process/user-process';
+import { NameSpace } from '../../const';
+import { offerData } from '../offer-data/offer-data';
+import { userProcess } from '../user-process/user-process';
 
 export const rootReducer = combineReducers({
   [NameSpace.Data]: offerData.reducer,
